arm: add SDIV/UDIV emulation for ARMv7 processors
The Virtualization Extensions introduce the requirement for an ARMv7-A
implementation to include SDIV and UDIV.  Any implementation of the
Virtualization Extensions must include the SDIV and UDIV instructions
in the Thumb and ARM instruction sets.

In an ARMv7-A implementation that does not include the Virtualization
Extensions, it is IMPLEMENTATION DEFINED whether:
 *  SDIV and UDIV are not implemented
 *  SDIV and UDIV are implemented only in the Thumb instruction set
 *  SDIV and UDIV are implemented in the Thumb and ARM instruction sets.

This patch adds a handler to trap and emulate unimplemented SDIV and
UDIV instructions in ARM and Thumb modes. Also some basic statistic is
exported via /proc/cpu/idiv_emulation

Signed-off-by: Vladimir Murzin <murzin.v at gmail.com>

Source: https://lists.linaro.org/pipermail/linaro-kernel/2012-September/002236.html